requirements.txt
protobuf==2.6.1
Twisted==14.0.2
txrudp==0.5.1
pystun==0.1.0
bitcointools==1.1.44
gnupg==2.0.2
pynacl==0.3.0
txrestapi==0.2
txws==0.9.1
python-libbitcoinclient==0.3.2
python-bitcoinlib==0.5.0
requests==2.7.0
pyopenssl==0.15.1
miniupnpc==1.9
service_identity
